+++ pakyow-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * [fixed] Unpredictable load order of backend aspects
     * Now loaded alphabetically on every os
+  * [fixed] App connection path is relative to to the app mount path
 
 # 1.0.1